pyproject.toml
[project]
name = "automobile-voice-assistant"
version = "0.1.0"
description = "Automobile Voice Assistant — Hindi/Hinglish telephony agent"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents",
    "livekit-plugins-openai",
    "livekit-plugins-silero",
    "livekit-plugins-turn-detector",
    "livekit-plugins-deepgram",
    "livekit-plugins-sarvam",
    "livekit-plugins-elevenlabs",
    "livekit-plugins-cartesia",
    "livekit-plugins-noise-cancellation",
    "python-dotenv",
    "asyncpg",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
